requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# GPT-OSS Integration
requests==2.31.0

# Future integrations (uncomment when ready)
# openai-harmony==0.1.0  # Uncomment when GPT-OSS integration is ready
# gpt-oss==0.1.0  # Uncomment when GPT-OSS integration is ready
